refactor(gateway): split lib.rs (5,998 lines) into route modules
Extract the monolithic lib.rs into focused modules:

- state.rs — AppState, HistogramBuckets, embedding provider cache
- middleware.rs — RateLimiter, auth/rate-limit/logging middleware, CORS
- sse.rs — SSE helpers (send_sse, CancelOnDrop)
- routes/chat.rs — chat, streaming, batch, WebSocket, OpenAI-compat
- routes/sessions.rs — session CRUD, fork, tags, import/export, insights
- routes/skills.rs — skill CRUD and usage stats
- routes/memories.rs — memory list/search/embed/delete
- routes/schedules.rs — schedule CRUD
- routes/user_model.rs — user trait CRUD
- routes/subagents.rs — subagent handlers
- routes/tools.rs — tool introspection and config
- routes/pairing.rs — DM pairing management
- routes/health.rs — health, MCP status, metrics, agent card
- routes/admin.rs — audit, analytics, cache, webhooks, templates,
  workflows, agent bus, eval, guardrails

lib.rs retains: router construction, shared helpers/pagination types,
embedded UI support, and all tests.

All 207 existing tests pass, zero clippy warnings.

Closes #273
